+++ src/sentry/api/utils.py
@@ -2,14 +2,12 @@ from __future__ import annotations
 
 import datetime
 import logging
-import re
 import sys
 import traceback
 from collections.abc import Generator, Mapping, MutableMapping
 from contextlib import contextmanager
 from datetime import timedelta
 from typing import Any, Literal, overload
-from urllib.parse import urlparse
 
 import sentry_sdk
 from django.conf import settings
@@ -31,6 +29,10 @@ from sentry.models.apikey import is_api_key_auth
 from sentry.models.apitoken import is_api_token_auth
 from sentry.models.organization import Organization
 from sentry.models.orgauthtoken import is_org_auth_token_auth
+from sentry.organizations.absolute_url import (  # noqa: F401 # XXX: for compatibility, remove after getsentry is updated
+    customer_domain_path,
+    generate_organization_url,
+)
 from sentry.organizations.services.organization import (
     RpcOrganization,
     RpcOrganizationMember,
@@ -271,25 +273,6 @@ def is_member_disabled_from_limit(
         return member.flags.member_limit__restricted
 
 
-def generate_organization_hostname(org_slug: str) -> str:
-    url_prefix_hostname: str = urlparse(options.get("system.url-prefix")).netloc
-    org_base_hostname_template: str = options.get("system.organization-base-hostname")
-    if not org_base_hostname_template:
-        return url_prefix_hostname
-    has_org_slug_placeholder = "{slug}" in org_base_hostname_template
-    if not has_org_slug_placeholder:
-        return url_prefix_hostname
-    org_hostname = org_base_hostname_template.replace("{slug}", org_slug)
-    return org_hostname
-
-
-def generate_organization_url(org_slug: str) -> str:
-    org_url_template: str = options.get("system.organization-url-template")
-    if not org_url_template:
-        return options.get("system.url-prefix")
-    return org_url_template.replace("{hostname}", generate_organization_hostname(org_slug))
-
-
 def generate_region_url(region_name: str | None = None) -> str:
     region_url_template: str | None = options.get("system.region-api-url-template")
     if region_name is None and SiloMode.get_current_mode() == SiloMode.REGION:
@@ -303,40 +286,6 @@ def generate_region_url(region_name: str | None = None) -> str:
     if not region_url_template or not region_name:
         return options.get("system.url-prefix")
     return region_url_template.replace("{region}", region_name)
-
-
-_path_patterns: list[tuple[re.Pattern[str], str]] = [
-    # /organizations/slug/section, but not /organizations/new
-    (re.compile(r"\/?organizations\/(?!new)[^\/]+\/(.*)"), r"/\1"),
-    # For /settings/:orgId/ -> /settings/organization/
-    (
-        re.compile(r"\/settings\/(?!account\/|!billing\/|projects\/|teams)[^\/]+\/?$"),
-        "/settings/organization/",
-    ),
-    # Move /settings/:orgId/:section -> /settings/:section
-    # but not /settings/organization or /settings/projects which is a new URL
-    (
-        re.compile(r"^\/?settings\/(?!account\/|billing\/|projects\/|teams)[^\/]+\/(.*)"),
-        r"/settings/\1",
-    ),
-    (re.compile(r"^\/?join-request\/[^\/]+\/?.*"), r"/join-request/"),
-    (re.compile(r"^\/?onboarding\/[^\/]+\/(.*)"), r"/onboarding/\1"),
-    (
-        re.compile(r"^\/?(?!settings)[^\/]+\/([^\/]+)\/getting-started\/(.*)"),
-        r"/getting-started/\1/\2",
-    ),
-]
-
-
-def customer_domain_path(path: str) -> str:
-    """
-    Server side companion to path normalizations found in withDomainRequired
-    """
-    for pattern, replacement in _path_patterns:
-        updated = pattern.sub(replacement, path)
-        if updated != path:
-            return updated
-    return path
 
 
 def method_dispatch(**dispatch_mapping):
