+++ src/sentry/integrations/github/integration.py
@@ -13,7 +13,6 @@ from django.utils.translation import gettext_lazy as _
 from rest_framework.request import Request
 
 from sentry import features, options
-from sentry.api.utils import generate_organization_url
 from sentry.constants import ObjectStatus
 from sentry.http import safe_urlopen, safe_urlread
 from sentry.identity.github import GitHubIdentityProvider, get_user_info
@@ -31,6 +30,7 @@ from sentry.integrations.models.organization_integration import OrganizationInte
 from sentry.integrations.services.repository import RpcRepository, repository_service
 from sentry.integrations.utils.code_mapping import RepoTree
 from sentry.models.repository import Repository
+from sentry.organizations.absolute_url import generate_organization_url
 from sentry.organizations.services.organization import RpcOrganizationSummary, organization_service
 from sentry.pipeline import Pipeline, PipelineView
 from sentry.shared_integrations.constants import ERR_INTERNAL, ERR_UNAUTHORIZED
