+++ src/sentry/middleware/customer_domain.py
@@ -11,7 +11,7 @@ from django.http.response import HttpResponseBase
 from django.urls import resolve, reverse
 
 from sentry import features
-from sentry.api.utils import generate_organization_url
+from sentry.organizations.absolute_url import generate_organization_url
 from sentry.organizations.services.organization import organization_service
 from sentry.types.region import subdomain_is_region
 from sentry.utils import auth
