+++ src/sentry/utils/auth.py
@@ -19,6 +19,7 @@ from sentry import options
 from sentry.models.organization import Organization
 from sentry.models.outbox import outbox_context
 from sentry.models.user import User
+from sentry.organizations.absolute_url import generate_organization_url
 from sentry.organizations.services.organization import RpcOrganization
 from sentry.users.services.user import RpcUser
 from sentry.users.services.user.service import user_service
@@ -190,8 +191,6 @@ def _get_login_redirect(request: HttpRequest, default: str | None = None) -> str
 
 
 def get_login_redirect(request: HttpRequest, default: str | None = None) -> str:
-    from sentry.api.utils import generate_organization_url
-
     login_redirect = _get_login_redirect(request, default)
     url_prefix = None
     if hasattr(request, "subdomain") and request.subdomain:
