+++ src/sentry/web/frontend/react_page.py
@@ -12,7 +12,7 @@ from django.urls import resolve
 from rest_framework.request import Request
 
 from sentry import features, options
-from sentry.api.utils import customer_domain_path, generate_organization_url
+from sentry.organizations.absolute_url import customer_domain_path, generate_organization_url
 from sentry.organizations.services.organization import organization_service
 from sentry.types.region import subdomain_is_region
 from sentry.users.services.user.model import RpcUser
